watch for Game End after all
We can't always rely on the length specified by UBJSON for the `raw`
element, as it will be zero for in-progress replays. So we still need to
stop once we hit Game End, even though the spec doesn't currently
guarantee this will be the last event in the stream.

Partially reverts 44c6cc1b9a7c691e2370ab8f962f3cf470bf83f0.